Divide test files with each attributes
