+++ tests/test_lowercase_only_ascii.py
@@ -1,0 +1,32 @@
+# coding:utf-8
+from base import TestBase
+import sublime
+import sys
+
+class TestLowercaseOnlyAscii(TestBase):
+    """Test of attributes 'lowercase_only_ascii'"""
+
+    # for debug
+    # def tearDown(self):
+    #     pass
+
+    lowercase_only_ascii_text = \
+"""
+
+<!-- MarkdownTOC uri_encoding=false {0} -->
+
+<!-- /MarkdownTOC -->
+
+# ПРИМЕР EXAMPLE
+"""
+    def test_lowercase_only_ascii_default(self):
+        toc_txt = self.commonSetup(self.lowercase_only_ascii_text.format('autolink=true'))
+        self.assert_In('- [ПРИМЕР EXAMPLE][ПРИМЕР-example]', toc_txt)
+
+    def test_lowercase_only_ascii_true(self):
+        toc_txt = self.commonSetup(self.lowercase_only_ascii_text.format('autolink=true lowercase_only_ascii=true'))
+        self.assert_In('- [ПРИМЕР EXAMPLE][ПРИМЕР-example]', toc_txt)
+
+    def test_lowercase_only_ascii_false(self):
+        toc_txt = self.commonSetup(self.lowercase_only_ascii_text.format('autolink=true lowercase_only_ascii=false'))
+        self.assert_In('- [ПРИМЕР EXAMPLE][пример-example]', toc_txt)
